+++ src/main/java/woowacourse/shoppingcart/dto/cartitem/CartItemRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.cartitem;
 
 public class CartItemRequest {
