+++ src/main/java/woowacourse/shoppingcart/dto/cartitem/CartItemResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.cartitem;
 
 import woowacourse.shoppingcart.domain.CartItem;
